Adding in matrix functionality
