+++ Maths_Engine/Compute/MatrixAddition.cs
@@ -21,24 +21,31 @@
  */
 
 using BH.oM.Base;
+
 using BH.oM.Base.Attributes;
-using BH.oM.Maths;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
+using MathNet.Numerics;
+using BH.oM.Maths;
 
 namespace BH.Engine.Maths
 {
-    public static partial class Query
+    public static partial class Compute
     {
-        [Description("Returns the number of coloumns in a matrix")]
-        [Input("Matrix", "A Matrix")]
-        [Output("Coloumns", "The number of coloumns in a matrix")]
-        public static int MatrixNumberOfColumns(this Matrix mat)
+
+        [Description("Computes the factorial of an integer accurately up until 170 where it will overflow.")]
+        [Input("Int", "Integer greater than or equal to 1.")]
+        [Output("Factorial", "Factorial of the integer")]
+        public static double MatrixAddtion(Matrix M1,Matrix M2)
         {
-            int cols = mat.Values.Count();
-            return cols;
+            if (BH.Engine.Maths.Query.IsAddableMatrix(M1,M2) == false)
+            {
+                BH.Engine.Base.Compute.RecordError("Number must be greater than or equal to 0.");
+            }
+            
+            return MathNet.Numerics.SpecialFunctions.Factorial(2);
         }
 
     }
